Rename interfaces to enums
