+++ src/drivers/generic/models/point.py
@@ -1,6 +1,6 @@
 from src import db
-from src.drivers.drivers import Drivers
-from src.drivers.generic.interfaces.point.points import GenericPointType
+from src.drivers.enums.drivers import Drivers
+from src.drivers.generic.enums.point.points import GenericPointType
 from src.models.point.model_point_mixin import PointMixinModel
