--- src/drivers/modbus/interfaces/network/rtu_network.py
@@ -1,27 +1,0 @@
-rtu_port = {
-    "/dev/ttyUSB0": "/dev/ttyUSB0",
-    "/dev/ttyUSB1": "/dev/ttyUSB1",
-    "/dev/ttyUSB2": "/dev/ttyUSB2",
-    "/dev/ttyUSB3": "/dev/ttyUSB3",
-    "/dev/ttyUSB4": "/dev/ttyUSB4",
-}
-
-rtu_parity = {
-    "even": "E",
-    "odd": "O",
-    "none": "N",
-}
-
-rtu_databits = {
-    "even": "E",
-    "odd": "O",
-    "none": "N",
-}
-
-rtu_baud_rate = {
-    "115200": 115200,
-    "57600": 57600,
-    "38400": 38400,
-    "19200": 19200,
-    "9600": 9600,
-}
