+++ src/drivers/modbus/models/network.py
@@ -1,8 +1,8 @@
 from sqlalchemy.orm import validates
 
 from src import db
-from src.drivers.drivers import Drivers
-from src.drivers.modbus.interfaces.network.network import ModbusType, ModbusRtuParity
+from src.drivers.enums.drivers import Drivers
+from src.drivers.modbus.enums.network.network import ModbusType, ModbusRtuParity
 from src.models.network.model_network_mixin import NetworkMixinModel
